List all actions AND moves in single query
With this change, the client should be able to display all valid actions
and moves and respond with a single action+move combo.

Also, caught bug where a player in a hall could make a suggestion
causing the gameplay to crash.  Now players in hallways cannot make
suggestions as the rules intend.