Added client directional awareness
The client is now able to make a decision as to whether or not it can
move in a certain direction. If a hallway is occupied by another player
then it is given the "occupied" = true property. This prevents other
players from moving there. Each client maintains their own copy of
player locations and updates the locations when a MOVE or NEWPLAYER
announcement comes in. I changed the gameboard layout to have ID tags
for everything to make it easier to move characters around.