Finished the Q_CARDS query type.
Reverting the A_ and Q_ strings.  Made the modal views impossible to
escape. Once you click on the button it is disabled and you have to do
what the dialog box asks you. Implemented displayCards, which will show
cards from the Q_CARDS query. The user can then pick the image and send
it back to the server.